Create books and categories slices in the features folder
