--- src/booksItems.js
@@ -1,3 +1,0 @@
-const booksItems = [];
-
-export default booksItems;
