+++ src/redux/features/categories/categoriesSlice.js
@@ -1,0 +1,15 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialState = [];
+
+const categoriesSlice = createSlice({
+  name: 'categories',
+  initialState,
+  reducers: {
+    checkStatus: () => ['Under construction'],
+  },
+});
+
+export const { checkStatus } = categoriesSlice.actions;
+
+export default categoriesSlice.reducer;
